+++ src/share/index.ts
@@ -1,2 +1,4 @@
 export { share, isShareSupported } from "./share";
 export type { SharePayload, ShareResult } from "./share";
+export { shareOrDownloadBlob } from "./share-or-download";
+export type { ShareOrDownloadOptions } from "./share-or-download";